requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio<4.0
asgiref==3.8.1
attrs==24.3.0
awsebcli==3.21.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
botocore==1.35.90
build==1.2.2.post1
cachetools==5.5.0
cement==2.10.14
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.9
emoji==2.14.0
eval_type_backport==0.2.2
fastapi==0.109.2
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.1.0
langchain-core==0.3.27
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-qdrant==0.2.0
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith>=0.1.0,<0.2.0
llama-index-core==0.12.8
llama-index-readers-file==0.4.1
llama-parse==0.3.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.19.1
nest-asyncio==1.6.0
nest_asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pathspec==0.10.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
posthog==3.7.4
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.5.3
pydantic-settings==2.7.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.1
tavily-python==0.5.0
tenacity==8.2.3
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.5.2
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.11
unstructured-client==0.28.1
urllib3>=1.26.5,<2
uvicorn==0.27.1
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wikipedia==1.4.0
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
qdrant-client==1.7.0
